+++ server/alerts.go
@@ -22,7 +22,7 @@ import (
 	"github.com/kevinburke/logrole/views"
 	"github.com/kevinburke/rest"
 	"github.com/kevinburke/rest/resterror"
-	twilio "github.com/kevinburke/twilio-go"
+	twilio "github.com/kevinburke/twilio-go/v2"
 )
 
 const alertPattern = `(?P<sid>NO[a-f0-9]{32})`