+++ views/alert.go
@@ -8,7 +8,7 @@ import (
 
 	types "github.com/kevinburke/go-types"
 	"github.com/kevinburke/logrole/config"
-	twilio "github.com/kevinburke/twilio-go"
+	twilio "github.com/kevinburke/twilio-go/v2"
 )
 
 type AlertPage struct {